+++ nirvana/tests/test_scripts.py
@@ -11,11 +11,12 @@ from nirvana.tests.util import remote_data_file, requires_remote
 
 
 @requires_remote
-def test_manga_axisym():
+def test_manga_axisym_gas():
     odir = remote_data_file('tests/8138')
     if os.path.isdir(odir):
         shutil.rmtree(odir)
 
+    # Test basic gas fit
     args = manga_axisym.parse_args(['8138', '12704', '--root', remote_data_file(), '--odir', odir,
                                     '-t', 'Gas', '--min_vel_snr', '5', '--min_sig_snr', '5',
                                     '--max_vel_err', '100', '--max_sig_err', '100',
@@ -41,6 +42,77 @@ def test_manga_axisym():
 
 
 @requires_remote
+def test_manga_axisym_gas_deconv():
+    odir = remote_data_file('tests/8138')
+    if os.path.isdir(odir):
+        shutil.rmtree(odir)
+
+    # Test basic gas deconvolve fit
+    args = manga_axisym.parse_args(['8138', '12704', '--root', remote_data_file(), '--odir', odir,
+                                    '-t', 'Gas', '--min_vel_snr', '5', '--min_sig_snr', '5',
+                                    '--max_vel_err', '100', '--max_sig_err', '100',
+                                    '--min_unmasked', '10', '--coherent', '--deconv',
+                                    '--skip_plots'])
+    manga_axisym.main(args)
+
+    main_output_file = os.path.join(odir, 'nirvana-manga-axisym-8138-12704-Gas.fits.gz')
+    assert os.path.isfile(main_output_file), 'Output file not created.'
+
+    with fits.open(main_output_file) as hdu:
+        assert len(hdu) == 24, 'Data model changed'
+        assert hdu['FITMETA'].data['RCHI2'] < 1.1, 'Fit dramatically changed'
+
+    shutil.rmtree(odir)
+
+
+@requires_remote
+def test_manga_axisym_str():
+    odir = remote_data_file('tests/8138')
+    if os.path.isdir(odir):
+        shutil.rmtree(odir)
+
+    # Test basic gas deconvolve fit
+    args = manga_axisym.parse_args(['8138', '12704', '--root', remote_data_file(), '--odir', odir,
+                                    '-t', 'Stars', '--min_vel_snr', '5', '--min_sig_snr', '5',
+                                    '--max_vel_err', '100', '--max_sig_err', '100',
+                                    '--min_unmasked', '10', '--coherent', '--skip_plots'])
+    manga_axisym.main(args)
+
+    main_output_file = os.path.join(odir, 'nirvana-manga-axisym-8138-12704-Stars.fits.gz')
+    assert os.path.isfile(main_output_file), 'Output file not created.'
+
+    with fits.open(main_output_file) as hdu:
+        assert len(hdu) == 24, 'Data model changed'
+        assert hdu['FITMETA'].data['RCHI2'] < 1.1, 'Fit dramatically changed'
+
+    shutil.rmtree(odir)
+
+
+@requires_remote
+def test_manga_axisym_str_deconv():
+    odir = remote_data_file('tests/8138')
+    if os.path.isdir(odir):
+        shutil.rmtree(odir)
+
+    # Test basic gas deconvolve fit
+    args = manga_axisym.parse_args(['8138', '12704', '--root', remote_data_file(), '--odir', odir,
+                                    '-t', 'Stars', '--min_vel_snr', '5', '--min_sig_snr', '5',
+                                    '--max_vel_err', '100', '--max_sig_err', '100',
+                                    '--min_unmasked', '10', '--coherent', '--deconv',
+                                    '--skip_plots'])
+    manga_axisym.main(args)
+
+    main_output_file = os.path.join(odir, 'nirvana-manga-axisym-8138-12704-Stars.fits.gz')
+    assert os.path.isfile(main_output_file), 'Output file not created.'
+
+    with fits.open(main_output_file) as hdu:
+        assert len(hdu) == 24, 'Data model changed'
+        assert hdu['FITMETA'].data['RCHI2'] < 1.1, 'Fit dramatically changed'
+
+    shutil.rmtree(odir)
+
+
+@requires_remote
 def test_manga_asymdrift():
     odir = remote_data_file('tests/8138')
     if os.path.isdir(odir):
@@ -53,6 +125,35 @@ def test_manga_asymdrift():
                                        '--str_min_vel_snr', '5', '--str_min_sig_snr', '10',
                                        '--str_max_vel_err', '100', '--str_max_sig_err', '100',
                                        '--min_unmasked', '10', '--coherent', '--skip_plots'])
+                                       #]) #, '--verbose', '2'])
+    manga_asymdrift.main(args)
+
+    main_output_file = os.path.join(odir, 'nirvana-manga-asymdrift-8138-12704.fits.gz')
+    assert os.path.isfile(main_output_file), 'Output file not created.'
+
+    with fits.open(main_output_file) as hdu:
+        assert len(hdu) == 66, 'Data model changed'
+        assert hdu['GAS_FITMETA'].data['RCHI2'] < 1.1, 'Fit dramatically changed'
+        assert hdu['STR_FITMETA'].data['RCHI2'] < 1.1, 'Fit dramatically changed'
+        assert hdu[0].header['RCHI2'] < 1.1, 'Fit dramatically changed'
+
+    shutil.rmtree(odir)
+
+
+@requires_remote
+def test_manga_asymdrift_dcnv():
+    odir = remote_data_file('tests/8138')
+    if os.path.isdir(odir):
+        shutil.rmtree(odir)
+
+    args = manga_asymdrift.parse_args(['8138', '12704', '--root', remote_data_file(),
+                                       '--odir', odir,
+                                       '--gas_min_vel_snr', '5', '--gas_min_sig_snr', '5',
+                                       '--gas_max_vel_err', '100', '--gas_max_sig_err', '100',
+                                       '--str_min_vel_snr', '5', '--str_min_sig_snr', '10',
+                                       '--str_max_vel_err', '100', '--str_max_sig_err', '100',
+                                       '--min_unmasked', '10', '--coherent', '--deconv',
+                                       '--skip_plots'])
                                        #]) #, '--verbose', '2'])
     manga_asymdrift.main(args)